Add an attribute for aliases
The whole point of this repo is to define a single name for each
service. However, some (like our friend cinder) have a large install
base already using old non-conforming names. While projects should not
have aliases in general, if we can list the ones that escaped before
we got this effort started, it can decrease the confusion for API
consumers.

Change-Id: I4b89c32c7eed714a6779ee28a670839e26356364