Add the rest of the services we know about
Starting from governance, add every service that has an API and
documentation for that API. service-types are taken from the project's
devstack plugin, as that's the only current location.

Some of the services had types that do not match the guidelines. An
attempt was made to define a new service-type and list the old one in
the aliases list. It's almost certain these will be the cause of some
amount of consternation. In several places making a change feels a bit
trivial, but they were all done so we could look at the entire corpus
and see whether alignment in those cases seems more valuable than not.

After the first round of feedback - some of the proposed changes were
incorporated here (when they seemed non-controversial) and some are
being moved to follow up patches.

storlets has no devstack plugin, so was omitted.

ceilometer was included, even though it's listed as deprecated, because
it still exists, thus still has users who call it by a name.

Several of the services do not have api-ref documentation. That should
be fixed at some point and become a requirement.

The name validation regex was expanded to accomdate ec2-api (there's really
no other sane name for that service) - but ensures nothing else can add
numbers.

Removed the validation regex for aliases. By their nature they are
non-conforming service types.

Change-Id: I8c204ac5664561f0c4b1856cd8e5beaf0230d246