Build a transformed JSON document for publication
The YAML format is intended for human interaction. Scripts also want
to consume this data, so we'd like to publish a machine-friendly format
for them. That format adds a version and a sha, so that people have
a way to trace the published data back to its source, as well as forward
and reverse mappings built from the data to simplify the questions
consumers will wind up querying from the data.

Change-Id: Ia62e01ca4380df0a36daaf76fe90e5f808ac7773