Reorder service-types.yaml, fix service_type regex
This is a followon to change I8c204ac5664561f0c4b1856cd8e5beaf0230d246

That change added the remaining known service types to
service-types.yaml, but the fields of the new entries were ordered
differently than the existing.  This change reorders them for
consistency.

The same change also updated the regex for service_type in the schema.
The intent was (partially) to ensure the name didn't end with a hyphen,
which was allowed by the previous regex.  However, in the process, the
new regex required a minimum of three characters in the name, whereas
the original only required two.  This change fixes the regex so
two-character names are allowed again (but the name still can't begin or
end with a hyphen).

Change-Id: Id310567f2597cca136f51ae2b2bb464d76d96741