Use 'clustering' as senlin's official service_type
Senlin has historically used `clustering` as its service_type.  We've
decided [1] to stop trying to retrofit existing historical services'
official service_types to conform to the naming guidelines, because it's
more painful than it's worth.

However, since at this point we've published sta with
`resource-cluster`, we now need to maintain that as an alias :(

[1] https://review.openstack.org/494241

Change-Id: Ic3c04c6d6ce98a56f333f8566d1cba9c65431313