Remove leading openstack/ from project name
The leading openstack/ is repetitive. There are zero chances anything
will make it into this file without a leading openstack/.

We're not actually using project for anything yet, but we want to use
it for both validation of documentation locations as well as being able
to send legacy-style microversion headers. Rather than removing the
openstack/ at processing time, remove it in the data.

Change-Id: I740af4211dcf142865f1fe70fec41ab5ae73d0e3