Only run json.dump once
Rather than dump twice, once for each file, dump once to a variable.
Then, while we've got it in memory, strip trailing spaces. It's not
required that json output have no trailing spaces, but gerrit shows them
as does vim when opening files to look at them. It's easy to rememdy,
and we're rolling out some file updates in the next patch anyway.

Change-Id: Id3d04655264480b202a58b93a813360003707508