Ensure git rev-parse output works in python3
basepython goes in testenv section, not in tox section. This means this
hasn't actually been running under python3. When running it in python3,
output from shell commands is bytes, not string, which means you get this:

  b'b92dbb8f25972dabdce851d34b97996b31f08b75' is not of type 'string'

Decoding the output is more correct in the python, and fixing the
tox.ini file means we won't regress.

Also fix an issue in doc/source/conf.py

Change-Id: I7d59a7ad9eda54e0347bdc3110cc8a95139f6e17