Generate standard api_reference
The api_reference for a service_type should usually be
'https://developer.openstack.org/api-ref/{service_type}/'.

With this change set, the source yaml may omit the api_reference field,
whereupon it will be generated according to the above format.

(At this time it is recommended, though not enforced, that the
api_reference field *should* be omitted if it conforms to the above
format.)

Change-Id: I860c1373c72765d4abad1ce8744c8174e5c64de0