Add historical aliases for openstacksdk names
python-openstacksdk has updated itself to drive the names of the service
objects it presents from the service-types-authority, using the official
service-type in the source code and docs but also registering each on
the Connection object by each of its known aliases. It was already
close, but in four cases it had chosen names that were more consistent
from an English point of view.

Add aliases for each of those.

Change-Id: I973b67043aa0bca305d7b10035567c175ac782b8