Add masakari and service-type instance-ha
The masakari project is an Official OpenStack project.

It has used the 'ha' service-type up until now, but that's a bit broad.
It was agreed that 'instance-ha' would be better. Add it to the registry.

The api-ref documentation is not currently published, but it does exist
in tree:

  http://git.openstack.org/cgit/openstack/masakari/tree/api-ref/source

The jobs to publish it have been added:

  https://review.openstack.org/#/c/534878

but as the api-ref publication jobs use service-types-authority to
determine the correct place to publish to, the api-ref docs won't exist
until we land this. Once we do and once a publication job runs, they
should be found at the appropriate place.

Change-Id: Icdd0b09060ddd71b2313a90c1d62d84c31cdbb8d